refactor(frontend)[OK-325] Rename complexToast context and move errors to classes directory
