+++ client/src/classes/errors/ApiError.ts
