+++ client/src/classes/errors/FabricObjectIdError.ts
