+++ client/src/components/common/board-permissions-select/BoardPermissionsSelect.tsx
@@ -6,8 +6,8 @@ import { BoardPermission } from "@/enums/BoardPermission";
 import { getPermissionLabel, getPermissionVariant } from "@/lib/userBoardsUtils";
 import logger from "@/lib/logger";
 import { toast } from "sonner";
-import { ApiError } from "@/errors/ApiError";
-import { complexToast } from "@/contexts/complexToast";
+import { ApiError } from "@/classes/errors/ApiError";
+import { complexToastContext } from "@/contexts/ComplexToastContext";
 import { ToastTypes } from "@/enums/ToastType";
 
 interface PermissionSelectProps {
@@ -38,7 +38,7 @@ const BoardPermissionsSelect: React.FC<PermissionSelectProps> = ({
     } catch (error: any) {
       logger.error(`Failed to change permission`, error);
       if (error instanceof ApiError) {
-        complexToast(ToastTypes.ERROR, error.messages, { duration: Infinity });
+        complexToastContext(ToastTypes.ERROR, error.messages, { duration: Infinity });
       } else {
         toast.error(error.message || "Failed to update permission");
       }
